% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpBites-package.R
\docType{package}
\name{ExpBites-package}
\alias{ExpBites}
\alias{ExpBites-package}
\title{ExpBites: Analyzing Human Exposure to Mosquito Biting}
\description{
Tools to analyse human and mosquito behavioral interactions and to compute exposure to mosquito bites estimates. Using behavioral data for human individuals and biting patterns for mosquitoes, you will be able to compute hourly exposure for bed net users and non-users, and summarize (e.g. proportion indoors and outdoors, proportion per time periods, and proportion prevented by bed nets) or visualize these dynamics across a 24-hour cycle.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nmoiroux/ExpBites}
  \item Report bugs at \url{https://github.com/Nmoiroux/ExpBites/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Moiroux \email{nicolas.moiroux@ird.fr} (\href{https://orcid.org/0000-0001-6755-6167}{ORCID})

}
\keyword{internal}
