% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gen_df_mosquito}
\alias{gen_df_mosquito}
\title{Generate a Simulated Dataset of Hourly Mosquito Biting Rates}
\usage{
gen_df_mosquito()
}
\value{
A data.frame with the following columns:
\describe{
\item{hour}{Hour of the day (0–23).}
\item{indoor_bites}{Simulated number of mosquito bites occurring indoors for each hour.}
\item{outdoor_bites}{Simulated number of mosquito bites occurring outdoors for each hour.}
}
}
\description{
Simulates the number of Anopheles mosquito bites occurring indoors and outdoors for each hour of the day.
Biting rates follow predefined lambda patterns that reflect typical mosquito activity (e.g., low during the day, peaking at night).
}
\examples{
df_bites <- gen_df_mosquito()
head(df_bites)

}
