% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calculate_Exp}
\alias{calculate_Exp}
\title{Calculate Hourly Mosquito Biting Exposure for Bednet Users and Non-Users}
\usage{
calculate_Exp(df, df_bites, protection = 0.92)
}
\arguments{
\item{df}{A dataframe with individual-level data, including columns:
\itemize{
\item \code{individual} – individual ID,
\item \code{hour} – hour of the day (0–23),
\item \code{state} – activity/location state (1 = outdoors, 2 = indoors awake,
3 = indoors asleep, 4 = indoors asleep under bednet),
\item \code{bednet_user} – logical indicator of bednet use (TRUE/FALSE).
}}

\item{df_bites}{A dataframe of biting rates by hour, with columns:
\itemize{
\item \code{hour} – hour of the day (0–23),
\item \code{indoor_bites} – number of bites indoors,
\item \code{outdoor_bites} – number of bites outdoors.
}}

\item{protection}{A numeric value (default = 0.92, according to Corbel et al. 2009 for Permanet 2 ITNs) indicating the proportion of bites prevented by bednet use.}
}
\value{
A dataframe with one row per hour (0–23) and the following columns:
\itemize{
\item \code{Bi_t} – indoor biting rate,
\item \code{Bo_t} – outdoor biting rate,
\item \code{N} – number of human individuals,
\item \code{Np} – number of human individuals that are net users,
\item \code{It} – proportion of people indoors (users and non-users),
\item \code{Eui} – mean exposure of non-users indoors,
\item \code{Euo} – mean exposure of non-users outdoors,
\item \code{Eu} – total mean exposure of non-users,
\item \code{p_in} – proportion of people (users) being indoors,
\item \code{p_net} – proportion of people (users) being asleep (under net),
\item \code{Epi} – mean exposure of users indoors (not under net),
\item \code{Epn} – mean exposure of users indoors under bednet (adjusted by protection),
\item \code{Epo} – mean exposure of users outdoors,
\item \code{Epp} – mean exposure prevented by bednet use,
\item \code{Ep} – total mean exposure of users.
}
}
\description{
Using a model of behavioral interactions, computes the hourly exposure to mosquito bites for both bednet users and non-users,
based on their location (indoors, outdoors, under a net) and mosquito biting rates.
For users, exposure is adjusted for the protective effect of bednets.
}
\details{
The mathematical model of behavioral interactions is an extension of the Killeen’s model (1) as previously described in Geissbühler et al. (2) and Moiroux et al. (3)

(1) Killeen, Gerry F., Japhet Kihonda, Edith Lyimo, Fred R. Oketch, Maya E. Kotas, Evan Mathenge, Joanna A. Schellenberg, Christian Lengeler, Thomas A. Smith, and Chris J. Drakeley. ‘Quantifying Behavioural Interactions between Humans and Mosquitoes: Evaluating the Protective Efficacy of Insecticidal Nets against Malaria Transmission in Rural Tanzania’. BMC Infectious Diseases 6 (2006): 161. https://doi.org/10.1186/1471-2334-6-161.

(2) Geissbühler, Yvonne, Prosper Chaki, Basiliana Emidi, Nicodemus J. Govella, Rudolf Shirima, Valeliana Mayagaya, Deo Mtasiwa, et al. ‘Interdependence of Domestic Malaria Prevention Measures and Mosquito-Human Interactions in Urban Dar Es Salaam, Tanzania’. Malaria Journal 6 (2007): 126. https://doi.org/10.1186/1475-2875-6-126.

(3) Moiroux, Nicolas, Georgia B. Damien, Marc Egrot, Armel Djenontin, Fabrice Chandre, Vincent Corbel, Gerry F. Killeen, and Cédric Pennetier. ‘Human Exposure to Early Morning Anopheles Funestus Biting Behavior and Personal Protection Provided by Long-Lasting Insecticidal Nets’. PloS One 9, no. 8 (2014): e104967. https://doi.org/10.1371/journal.pone.0104967.

Reference for the \code{protection} parameter:

(4) Corbel, Vincent, Joseph Chabi, Roch K. Dabiré, Josiane Etang, Philippe Nwane, Olivier Pigeon, Martin Akogbeto, and Jean-Marc Hougard. ‘Field Efficacy of a New Mosaic Long-Lasting Mosquito Net (PermaNet 3.0) against Pyrethroid-Resistant Malaria Vectors: A Multi Centre Study in Western and Central Africa’. Malaria Journal 9 (27 April 2010): 113. https://doi.org/10.1186/1475-2875-9-113.
}
\examples{
# generate fake data 
df <- gen_df_human(n_individuals = 1, hours = c(0:9,17:23))
df_bites <- gen_df_mosquito() 
# calculate mean hourly exposure to bites
exposure_results <- calculate_Exp(df, df_bites)

}
